fix after review

+++ models/event.go
@@ -6,7 +6,7 @@ import (
 )
 
 type Event struct {
-	UserID               int             `json:"user_id" validate:"required,gt=0"`
+	UserID               int64           `json:"user_id" validate:"required,gt=0"`
 	EventActionTimestamp time.Time       `json:"event_action_timestamp" validate:"required"`
 	Action               string          `json:"action" validate:"required,oneof=created updated deleted viewed"`
 	Metadata             json.RawMessage `json:"metadata" validate:"required,json"`
